Automated g4 rollback of 16ed1890
*** Reason for rollback ***

This CL is a "rollforward" of CL bf4e59b4 plus a patch to Expect. Previously only a AssumptionViolatedException would be re-thrown if no failures had been gathered by Expect up to that point. This has been changed to apply to all Throwables.

Now if an exception is thrown by a test and no failures have been gathered by Expect, the exception is simply re-thrown. Only if previous failures have been gathered will the new exception be appended to the list of failures which will then be thrown as before as an AssertionError.

*** Original change description ***

Automated g4 rollback of changelist bf4e59b4.

*** Reason for rollback ***

Breaks expected-exception tests that expect the original exception to be propagated.

*** Original change description ***

Change Expect to catch an exception thrown during the evaluation of a Statement and treat is as a failure.

Change ExpectationGatherer::toString to use stripTruthStackTrace on any failure causes.

Background:
Prior to this change, if an exception was thrown in a test, any previous failures gathered...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165646094